# swagger.yaml
openapi: 3.0.0
info:
  title: EduRail API
  version: 1.0.0
  description: API Website Edukasi Perkeretaapian EduRail
servers:
  - url: http://localhost:5000/api
...
paths:
...
  /semboyan:
    get:
      tags:
        - Semboyan
      summary: Ambil semua data semboyan
      responses:
        '200':
          description: Success

    post:
      tags:
        - Semboyan
      summary: Tambah semboyan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Semboyan'

      responses:
        '201':
          description: Data berhasil ditambahkan
...
  /kereta:
    get:
      tags:
        - Kereta
      summary: Ambil semua data kereta
      responses:
        '200':
          description: Success

    post:
      tags:
        - Kereta
      summary: Tambah data kereta
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Kereta'

      responses:
        '201':
          description: Data berhasil ditambahkan
...
  /artikel:
    get:
      tags:
        - Artikel
      summary: Ambil semua artikel
      responses:
        '200':
          description: Success

    post:
      tags:
        - Artikel
      summary: Tambah artikel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Artikel'

      responses:
        '201':
          description: Berhasil ditambahkan
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Semboyan:
      type: object
      properties:
        nama:
          type: string
          example: Semboyan 35
        arti:
          type: string
          example: Berhenti
        gambar:
          type: string
          example: https://example.com/semboyan35.png
        deskripsi:
          type: string
          example: Kereta harus berhenti
        warna:
          type: string
          example: Merah
    Kereta:
      type: object
      properties:
        nama:
          type: string
          example: Argo Bromo Anggrek
        jenis:
          type: string
          example: Kereta Penumpang
        gambar:
          type: string
          example: https://example.com/argo.jpg
        deskripsi:
          type: string
          example: Kereta api kelas eksekutif
    Artikel:
      type: object
      properties:
        judul:
          type: string
          example: Mengenal Semboyan 35
        isi:
          type: string
          example: Semboyan 35 digunakan untuk...
        tanggal:
          type: string
          format: date
          example: 2026-06-10